Modified colors on landing page for readability
This commit aims to improve the accessibility score of the lighthouse audit.
The change brings the score up from 81 to 92.

Specifically, this commit does 2 things. First, it changes certain colors so
that they have a contrast ratio of at least 4.5:1. It modifies the background
color of the 'Submit' button in the search bar and the main color of the
'View' details buttons. Second, it changes 2 viewports to be user scalable.